STM8 port: Add RTOS port for IAR EWSTM8 compiler. * Split assembler routines into -cosmic.s and -iar.s. * Split Cosmic and IAR specific code using __CSMC__ and __IAR_SYSTEMS_ICC__. * Context-switch virtual registers on IAR. * Add detailed description of IAR context-switch scheme. * Add sample IAR project. * Add support for IAR to STM8S peripheral driver modules. * Fix EWSTM8 compiler warnings.

+++ tests/timer1.c
@@ -139,7 +139,7 @@ uint32_t test_start (void)
         /* Check that time has advanced by exactly 1 tick */
         if ((end_time - start_time) != 1)
         {
-            ATOMLOG (_STR("Tick1:%d\n"), (end_time-start_time));
+            ATOMLOG (_STR("Tick1:%d\n"), (int)(end_time-start_time));
             failures++;
         }
     }
@@ -175,7 +175,7 @@ uint32_t test_start (void)
         /* Check that time has advanced by exactly 2 ticks */
         if ((end_time - start_time) != 2)
         {
-            ATOMLOG (_STR("Tick2:%d\n"), (end_time-start_time));
+            ATOMLOG (_STR("Tick2:%d\n"), (int)(end_time-start_time));
             failures++;
         }
     }
@@ -211,7 +211,7 @@ uint32_t test_start (void)
         /* Check that time has advanced by exactly 500 ticks */
         if ((end_time - start_time) != 500)
         {
-            ATOMLOG (_STR("Tick500:%d\n"), (end_time-start_time));
+            ATOMLOG (_STR("Tick500:%d\n"), (int)(end_time-start_time));
             failures++;
         }
     }
